Adapt toolbars to editors display
