+++ newIDE/app/src/SceneEditor/SwipeableDrawerEditorsDisplay/BottomToolbar.js
