+++ newIDE/app/src/SceneEditor/SwipeableDrawerEditorsDisplay/UseSwipeGesture.js
